solution for the last 'hard' problem
In this solution, an AppState is created to pass data between the two views. When ImmersiveView is initialized, it sets the values for the ballEntity and the initialPosition. We also add a button to ContentView that calls a custom method, ResetBall. This func grabs appState, which is stored in the Environment, and applies initialValue to the ballEntity's transform.translate property.